All assets types moved to asset folder
