+++ Jeu/Src/Common/Types/Assets/AnimatedAsset.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../ComponentsManagement/Animation/Animation.h"
+#include "../../../ComponentsManagement/Animation/Animation.h"
 #include <SDL2/SDL.h>
 #include <vector>
 #include <string>
